Title: 바이러스 검사, Time: 348ms, Memory: 4MB, Status: Passed - Codetree

+++ 250403/바이러스 검사/virus-detector.cpp
@@ -11,7 +11,7 @@ int main() {
     }
     int leaderCnt=0, memberCnt=0;
     cin >> leaderCnt >> memberCnt;
-    int cnt = 0;
+    long long cnt = 0;
     for(int i=0;i<n;i++){
         int num = v[i];
         if(num-leaderCnt<0){
